+++ packages/core/src/types.ts
@@ -3,7 +3,41 @@
 import { InterpreterFrom } from 'xstate'
 
 import { AuthMachine } from './machines'
+import { StorageGetter, StorageSetter } from './storage'
 
+export interface AuthOptions {
+  /** Time interval until token refreshes, in seconds */
+  refreshIntervalTime?: number
+  /**
+   * Define a way to get information about the refresh token and its exipration date.
+   * @default web */
+  clientStorageType?: ClientStorageType
+  /** Object where the refresh token will be persisted and read locally.
+   *
+   * Recommended values:
+   * - `'web'` and `'cookies'`: no value is required
+   * - `'react-native'`: `import Storage from @react-native-async-storage/async-storage`
+   * - `'cookies'`: `localStorage`
+   * - `'custom'`: an object that defines the following methods:
+   *     - `setItem` or `setItemAsync`
+   *     - `getItem` or `getItemAsync`
+   *     - `removeItem`
+   * - `'capacitor``: `import { Storage } from @capacitor/storage`
+   * - `'expo-secure-store'`: `import * as SecureStore from 'expo-secure-store'`
+   */
+  clientStorage?: ClientStorage
+  /**
+   * @deprecated Use clientStorage / clientStorageType instead */
+  clientStorageGetter?: StorageGetter
+  /**
+   * Define a way to set information about the refresh token and its exipration date.
+   * @deprecated  Use clientStorage / clientStorageType instead */
+  clientStorageSetter?: StorageSetter
+  /** When set to true, will automatically refresh token before it expires */
+  autoRefreshToken?: boolean
+  /** When set to true, will parse the url on startup to check if it contains a refresh token to start the session with */
+  autoSignIn?: boolean
+}
 // TODO import generated typings from 'hasura-auth'
 export type AuthInterpreter = InterpreterFrom<AuthMachine>
 interface RegistrationOptions {
@@ -134,3 +168,38 @@ export interface JWTClaims {
   iat?: number
   'https://hasura.io/jwt/claims': JWTHasuraClaims
 }
+
+export interface ClientStorage {
+  // custom
+  // localStorage
+  // AsyncStorage
+  // https://react-native-community.github.io/async-storage/docs/usage
+  setItem?: (_key: string, _value: string) => void
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  getItem?: (key: string) => any
+  removeItem?: (key: string) => void
+
+  // capacitor
+  set?: (options: { key: string; value: string }) => void
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  get?: (options: { key: string }) => any
+  remove?: (options: { key: string }) => void
+
+  // expo-secure-storage
+  setItemAsync?: (key: string, value: string) => void
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  getItemAsync?: (key: string) => any
+  deleteItemAsync?: (key: string) => void
+  customGet?: (key: string) => Promise<string | null> | string | null
+  customSet?: (key: string, value: string | null) => Promise<void> | void
+}
+
+// supported client storage types
+export type ClientStorageType =
+  | 'capacitor'
+  | 'custom'
+  | 'expo-secure-storage'
+  | 'localStorage'
+  | 'react-native'
+  | 'web'
+  | 'cookie'